Delete function modification , eslint check

+++ build/delete.js
@@ -11,10 +11,12 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const fs_1 = require("fs");
 const index_1 = require("./index");
-const Delete = (imageName) => __awaiter(void 0, void 0, void 0, function* () {
+const Delete = (imageName, exceptionimag) => __awaiter(void 0, void 0, void 0, function* () {
     const Files = yield fs_1.promises.readdir(index_1.thumbpath);
     Files.forEach((file) => __awaiter(void 0, void 0, void 0, function* () {
-        if (file.includes(imageName)) {
+        if (file.includes(imageName) && file !== `${exceptionimag}.jpg`) {
+            console.log(file);
+            console.log(exceptionimag);
             yield fs_1.promises.unlink(`${index_1.thumbpath}${file}`);
         }
     }));
